Debugged!
Ha ha! Though I don't get why the error existed.
I used console.log to check outputs.

and for the calculate_rent section it was outputting a string variable.

This led the total variable to become a string variable when it was
adding everything up.

And then the output percentage just stayed at 0%

I used Math.round to force it to become a value - but I didn't need to
do that for other variables.

Weirdness.